+++ include/geometry/poly_double.h
@@ -33,9 +33,9 @@ class Bound;
 class PolyDouble
 {
 	public:
-		size_t numVertices() const { return size_; }
+		int numVertices() const { return size_; }
 		bool empty() const { return size_ == 0; }
-		Vec3d operator[](size_t index) const { return vertices_[index]; }
+		Vec3d operator[](int index) const { return vertices_[index]; }
 		void addVertex(const Vec3d &vertex) { vertices_[size_++] = vertex; }
 		std::string print() const;
 		struct ClipResult;
@@ -47,7 +47,7 @@ class PolyDouble
 		static PolyDouble::ClipResult planeClip(Logger &logger, double pos, const ClipPlane &clip_plane, const PolyDouble &poly);
 		static Bound getBound(const PolyDouble &poly);
 		std::array<Vec3d, 10> vertices_; //Limited to triangles + 6 clipping planes cuts, or to quads + 6 clipping planes (total 10 edges/vertices max)
-		size_t size_ = 0;
+		int size_ = 0;
 };
 
 struct PolyDouble::ClipResult
